Create generate_conditions_v2.R
this is an alternative of what's in generate_conditions that does the (expected?) behavior of splitting the whole domain into the boundaries indicated by the entirety of breakpoints -- treating all detected breaks as a singular hypothesis. the current implementation treats each row individually, and therefore can only split the domain into a max of four boxes at once -- the total set of fits may correspond to more than four boxes, but there is overlap in the data used.